clean

+++ lib/Models.py
@@ -7,6 +7,6 @@ def LoadLLM(name):
     if tokenizer.pad_token is None:
         tokenizer.pad_token = tokenizer.eos_token
         
-    model = AutoModelForCausalLM.from_pretrained(name, device_map='cuda', torch_dtype=torch.float16)
+    model = AutoModelForCausalLM.from_pretrained(name, device_map='cuda')
 
     return model, tokenizer
